Refactor oauth revoke and oauth metadata into separate endpoint files
